requirements.txt
# Core Django
Django==4.2.10
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-widget-tweaks==1.5.0
Pillow==11.2.1

# Authentication and Authorization
django-allauth==0.58.2
django-auditlog==2.3.0

# Database and Configuration
django-environ==0.11.2
dj-database-url==2.1.0
python-decouple==3.8

# Data Management
django-tables2==2.7.0
django-filter==23.5

# Development Tools
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Dependencias adicionales
pillow==11.2.1

